+++ cnDefaultComponents.js
@@ -8,8 +8,8 @@ class CNCheckboxIn extends HTMLElement {
 		const checkedByDefault = this.getAttribute('checked') === '';
 		this.shadowRoot.innerHTML = `
 		${styles}	
-		<label for="cn-cbx" class="cn-cbx-label">
-        	<input type="checkbox" class="cn-cbx" id="cn-cbx"
+		<label  class="cn-cbx-label">
+        	<input type="checkbox" id="cn-cbx"
 			${checkedByDefault ? 'checked' : ''}/>
 			<span class="cn-cbx-checkmark"><slot></slot></span>
 		</label>
@@ -27,6 +27,58 @@ class CNCheckboxIn extends HTMLElement {
 
 customElements.define('cn-checkbox-in', CNCheckboxIn);
 
+class CNRadioButtonIn extends HTMLElement {
+	constructor() {
+		super();
+		this.attachShadow({ mode: 'open' });
+	}
+
+	renderRadioButtons(data) {
+		if (!data || data.length == 0) {
+			return;
+		}
+
+		let renderedButtons = '';
+		data.forEach(({ value, label, checked }) => {
+			renderedButtons += this.createRadioButton(value, label, checked);
+		});
+		this.shadowRoot.innerHTML = `
+		${styles}
+		${renderedButtons}
+		`;
+	}
+
+	createRadioButton(value, label, checked) {
+		return `
+		<label class="cn-radio-wrapper cn-radio-wrapper-before">
+        	<span class="cn-radio-input">
+				<input type="radio" value="${value}" ${
+			checked ? 'checked' : ''
+		} name="cn-radio">
+				<span class="cn-radio-control"></span>
+			</span>
+			<span class="cn-radio-label">${label}</span>
+		</label>
+		`;
+	}
+
+	getSelected() {
+		const radioButtons = this.shadowRoot.querySelectorAll('[name=cn-radio]');
+		const selected = Array.from(radioButtons).find(btn => btn.checked);
+		if (!selected) {
+			return null;
+		}
+		return selected.value;
+	}
+
+	unselectAll() {
+		const radioButtons = this.shadowRoot.querySelectorAll('[name=cn-radio]');
+		Array.from(radioButtons).forEach(btn => (btn.checked = false));
+	}
+}
+
+customElements.define('cn-radio-in', CNRadioButtonIn);
+
 const styles = `
 	<style>
 			* {
@@ -37,7 +89,13 @@ const styles = `
 			--cn-medium-grey: #474747;
 			--cn-dark-grey: #2c2c2c;
 			font-family: sans-serif;
-    		font-size: 1.1rem;
+    		font-size: 1rem;
+		}
+
+		*,
+		*:before,
+		*:after {
+			box-sizing: border-box;
 		}
 
 
@@ -84,19 +142,18 @@ const styles = `
 			outline: none;
 		}
 
-		.cn-cbx-label{
+		.cn-cbx-label {
 			display: inline-flex;
 			cursor: pointer;
 			position: relative;
 			 align-items: center;
 		}
 
-		.cn-cbx-label > span{
-			color: #fffff;
+		.cn-cbx-label > span {
 			padding: 0.5rem 0.25rem;
 		}
 
-		.cn-cbx-label > input{
+		.cn-cbx-label > input {
 			-webkit-appearance: none;
 			-moz-appearance: none;
 			-o-appearance: none;
@@ -111,11 +168,11 @@ const styles = `
 			
 		}
 
-		.cn-cbx-label > input:checked{
+		.cn-cbx-label > input:checked {
 			background-color: var(--cn-blue);
 		}
 		
-		.cn-cbx-label > input:checked + span::before{
+		.cn-cbx-label > input:checked + span::before {
 			content: '\\2713';
 			display: block;
 			text-align: center;
@@ -125,9 +182,76 @@ const styles = `
 			top: 0.4rem;
 		}
 
-		.cn-cbx-label > input:active{
+		.cn-cbx-label > input:active {
 			background-color: #ccc
 		}
+
+		.cn-radio-wrapper {
+			display: grid;
+			grid-template-columns: -webkit-min-content auto;
+			grid-template-columns: min-content auto;
+			grid-gap: 0.2em;
+			font-size: 2.25rem;
+			padding: 0.25rem 0;
+		}
+
+		.cn-radio-wrapper:focus-within .cn-radio-label {
+			color: var(--cn-blue-light)
+			
+		}
+
+		.cn-radio-label {
+			line-height: 1;
+			transition: 180ms all ease-in-out;
+		}
+
+		.cn-radio-input {
+			display: flex;
+		}
+
+		.cn-radio-input input {
+			opacity: 0;
+			width: 0;
+			height: 0;
+		}
+
+		.cn-radio-input input:focus + .cn-radio-control {
+			box-shadow: 0 0 0 0.05em #fff, 0 0 0.15em 0.1em currentColor;
+		}
+
+		.radio-gradient input:checked + .cn-radio-control {
+			background: radial-gradient(currentcolor 50%, rgba(255, 0, 0, 0) 51%);
+		}
+
+		.cn-radio-wrapper-before .cn-radio-control {
+			display: grid;
+			place-items: center;
+		}
+
+		.cn-radio-wrapper-before input + .cn-radio-control::before {
+			content: "";
+			width: 0.5em;
+			height: 0.5em;
+			box-shadow: inset 0.5em 0.5em currentColor;
+			border-radius: 50%;
+			transition: 180ms transform ease-in-out;
+			transform: scale(0);
+		}
+
+		.cn-radio-wrapper-before input:checked + .cn-radio-control::before {
+			transform: scale(1);
+		}
+
+		.cn-radio-control {
+			display: block;
+			width: 1em;
+			height: 1em;
+			border-radius: 50%;
+			border: 0.1em solid currentColor;
+			
+			color: var(--cn-blue)
+		}
+		
 	
 		</style>
 	`;